ambient: refactor controller implementation a bit
* Consistently use resource key as `network/foo`. Use string instead of
  NamespaceName and remove logic that assumes other key formats
* Add a helper to avoid so many type switches
* Move out authz rules to its own file; no changes, just moving
* A few other nit fixes as followup from PR
